requirements.txt
# ─── Core ────────────────────────────────────────────────────────────────────
numpy>=1.24
pandas>=2.0
Pillow>=10.0
matplotlib>=3.7
tqdm>=4.65

# ─── Image processing ────────────────────────────────────────────────────────
opencv-python-headless>=4.8
scikit-image>=0.21

# ─── Deep learning ───────────────────────────────────────────────────────────
torch>=2.0
torchvision>=0.15
pytorch-lightning>=2.0
timm>=0.9

# ─── Augmentation ────────────────────────────────────────────────────────────
albumentations>=1.3

# ─── Experiment tracking ─────────────────────────────────────────────────────
wandb>=0.16

# ─── ML utilities ────────────────────────────────────────────────────────────
scikit-learn>=1.3

# ─── FSCNN repo (install manually — not on PyPI) ──────────────────────────────
# git clone https://github.com/aimi-bonn/hand-segmentation.git
# sys.path.append("<repo_root>/FSCNN")